20_RealTime_Llama_Kafka/requirements.txt
streamlit
kafka-python
llama-cpp-python
pandas